refactor(agents): consolidate request routing
