+++ packages/agents/src/lifecycle/durable-object-lifecycle.ts
@@ -1,4 +1,4 @@
-import { DurableObject, env as defaultEnv } from "cloudflare:workers";
+import { DurableObject } from "cloudflare:workers";
 import { nanoid } from "nanoid";
 
 import {
@@ -89,201 +89,14 @@ function closeQuietly(ws: WebSocket, code: number, reason: string): void {
   }
 }
 
-// Cache binding discovery per environment object.
-const namespaceMapCache = new WeakMap<
-  object,
-  Record<string, DurableObjectNamespace>
->();
-
-// Maps kebab-case namespace -> original env binding name (e.g. "my-agent" -> "MyAgent")
-const bindingNameCache = new WeakMap<object, Record<string, string>>();
-
-const DEFAULT_ROUTING_RETRY_OPTIONS = {
-  maxAttempts: 3,
-  baseDelayMs: 100,
-  maxDelayMs: 800
-};
-
-/** Details supplied before a routed Durable Object request is retried. */
-export interface RoutingRetryEvent {
-  /** The retryable platform error. */
-  error: unknown;
-  /** The next attempt number. */
-  attempt: number;
-  /** Maximum attempts, including the initial request. */
-  maxAttempts: number;
-  /** Delay before the next attempt. */
-  delayMs: number;
-  /** Named Durable Object instance. */
-  name: string;
-  /** Matched Durable Object binding name, when known. */
-  className?: string;
-}
-
-/** Retry policy for routed Durable Object infrastructure failures. */
-export interface RoutingRetryOptions {
-  /** Max number of attempts, including the first. Default: 3 */
-  maxAttempts?: number;
-  /** Base delay in ms for exponential backoff. Default: 100 */
-  baseDelayMs?: number;
-  /** Max delay cap in ms. Default: 800 */
-  maxDelayMs?: number;
-  /** Optional callback invoked before each retry delay. */
-  onRetry?: (event: RoutingRetryEvent) => void | Promise<void>;
-}
-
-interface ResolvedRoutingRetryOptions extends Required<
-  Omit<RoutingRetryOptions, "onRetry">
-> {
-  onRetry?: RoutingRetryOptions["onRetry"];
-}
-
-interface RoutingRetryContext {
-  name: string;
-  className?: string;
-}
-
-/** @internal Convert public placement options for namespace.get(). */
-export function durableObjectGetOptions(
-  options: { locationHint?: DurableObjectLocationHint } | undefined
-) {
-  return options?.locationHint
-    ? { locationHint: options.locationHint }
-    : undefined;
-}
-
-function validatePositiveInteger(value: number, name: string): void {
-  if (!Number.isFinite(value) || value < 1) {
-    throw new Error(`${name} must be >= 1`);
-  }
-  if (!Number.isInteger(value)) {
-    throw new Error(`${name} must be an integer`);
-  }
-}
-
-function validatePositiveNumber(value: number, name: string): void {
-  if (!Number.isFinite(value) || value <= 0) {
-    throw new Error(`${name} must be > 0`);
-  }
-}
-
-function resolveRoutingRetryOptions(
-  options: false | RoutingRetryOptions | undefined
-): ResolvedRoutingRetryOptions | null {
-  if (options === false) return null;
-
-  const resolved = {
-    maxAttempts:
-      options?.maxAttempts ?? DEFAULT_ROUTING_RETRY_OPTIONS.maxAttempts,
-    baseDelayMs:
-      options?.baseDelayMs ?? DEFAULT_ROUTING_RETRY_OPTIONS.baseDelayMs,
-    maxDelayMs: options?.maxDelayMs ?? DEFAULT_ROUTING_RETRY_OPTIONS.maxDelayMs,
-    onRetry: options?.onRetry
-  };
-
-  validatePositiveInteger(resolved.maxAttempts, "routingRetry.maxAttempts");
-  validatePositiveNumber(resolved.baseDelayMs, "routingRetry.baseDelayMs");
-  validatePositiveNumber(resolved.maxDelayMs, "routingRetry.maxDelayMs");
-  if (resolved.baseDelayMs > resolved.maxDelayMs) {
-    throw new Error("routingRetry.baseDelayMs must be <= maxDelayMs");
-  }
-
-  return resolved;
-}
-
-function isRetryableDurableObjectError(error: unknown): boolean {
-  if (typeof error !== "object" || error === null) return false;
-  const typed = error as { retryable?: boolean; overloaded?: boolean };
-  return typed.retryable === true && typed.overloaded !== true;
-}
-
-function routingRetryDelayMs(
-  attempt: number,
-  options: ResolvedRoutingRetryOptions
-): number {
-  const upperBoundMs = Math.min(
-    options.maxDelayMs,
-    options.baseDelayMs * 2 ** (attempt - 1)
-  );
-  return Math.floor(Math.random() * upperBoundMs);
-}
-
-/** @internal Retry a transient Durable Object routing operation. */
-export async function retryDurableObjectOperation<T>(
-  operation: () => Promise<T>,
-  context: RoutingRetryContext,
-  retryOptions: false | RoutingRetryOptions | undefined
-): Promise<T> {
-  const resolved = resolveRoutingRetryOptions(retryOptions);
-  if (!resolved) return await operation();
-
-  let attempt = 1;
-  while (true) {
-    try {
-      return await operation();
-    } catch (error) {
-      const nextAttempt = attempt + 1;
-      if (
-        nextAttempt > resolved.maxAttempts ||
-        !isRetryableDurableObjectError(error)
-      ) {
-        throw error;
-      }
-
-      const delayMs = routingRetryDelayMs(attempt, resolved);
-      try {
-        await resolved.onRetry?.({
-          error,
-          attempt,
-          maxAttempts: resolved.maxAttempts,
-          delayMs,
-          name: context.name,
-          className: context.className
-        });
-      } catch (callbackError) {
-        console.warn(
-          "Durable Object routing retry callback failed:",
-          callbackError
-        );
-      }
-      await new Promise((resolve) => setTimeout(resolve, delayMs));
-      attempt = nextAttempt;
-    }
-  }
-}
-
-/**
- * Encode props for the internal lifecycle props header.
- *
- * The value travels in an HTTP header, so it must be ASCII-safe. Raw
- * `JSON.stringify` output can contain non-ASCII characters (e.g. accented
- * names like "Usuário"), which makes workerd emit a "header value contains
- * non-ASCII characters" warning and throws in browser fetch implementations.
- * We UTF-8 encode the JSON and base64 it so the header is always ASCII.
- */
-function encodeProps(props: unknown): string {
-  const bytes = new TextEncoder().encode(JSON.stringify(props));
-  let binary = "";
-  for (let i = 0; i < bytes.length; i++) {
-    binary += String.fromCharCode(bytes[i]);
-  }
-  return btoa(binary);
-}
-
 function mutableRequest(request: Request): Request {
   return new Request(request);
-}
-
-function cloneRequestForFetch(request: Request): Request {
-  // SAFETY: Workers Types v5 preserves an unconstrained `cf` generic from
-  // clone(). Cloning does not change the runtime Request representation.
-  return request.clone() as Request;
 }
 
 /**
  * Decode props from the internal lifecycle props header.
  *
- * Handles both the base64 encoding produced by `encodeProps` and, for
+ * Handles both base64-encoded lifecycle props and, for
  * backwards compatibility with stubs/requests created by older versions,
  * raw JSON. Base64 never starts with `{` or `[`, so a leading brace/bracket
  * unambiguously identifies the legacy raw-JSON form.
@@ -299,248 +112,6 @@ function decodeProps(header: string): unknown {
     bytes[i] = binary.charCodeAt(i);
   }
   return JSON.parse(new TextDecoder().decode(bytes));
-}
-
-function camelCaseToKebabCase(str: string): string {
-  // If string is all uppercase, convert to lowercase
-  if (str === str.toUpperCase() && str !== str.toLowerCase()) {
-    return str.toLowerCase().replace(/_/g, "-");
-  }
-
-  // Otherwise handle camelCase to kebab-case
-  let kebabified = str.replace(
-    /[A-Z]/g,
-    (letter) => `-${letter.toLowerCase()}`
-  );
-  kebabified = kebabified.startsWith("-") ? kebabified.slice(1) : kebabified;
-  // Convert any remaining underscores to hyphens and remove trailing -'s
-  return kebabified.replace(/_/g, "-").replace(/-$/, "");
-}
-/** Named Durable Object route matched in the outer Worker. */
-export interface DurableObjectRouteMatch<Env = Cloudflare.Env> {
-  /** The Durable Object class name / environment binding name. */
-  className: Extract<keyof Env, string>;
-  /** The named Durable Object instance extracted from the URL. */
-  name: string;
-}
-
-/** Options supported by `routeDurableObjectRequest`. */
-export interface DurableObjectRouteOptions<
-  Env = Cloudflare.Env,
-  Props = Record<string, unknown>
-> {
-  prefix?: string;
-  jurisdiction?: DurableObjectJurisdiction;
-  locationHint?: DurableObjectLocationHint;
-  props?: Props;
-  /**
-   * Whether to enable CORS for matched routes.
-   *
-   * When `true`, uses default permissive CORS headers:
-   * - Access-Control-Allow-Origin: *
-   * - Access-Control-Allow-Methods: GET, POST, HEAD, OPTIONS
-   * - Access-Control-Allow-Headers: *
-   * - Access-Control-Max-Age: 86400
-   *
-   * For credentialed requests, pass explicit headers with a specific origin:
-   * ```ts
-   * cors: {
-   *   "Access-Control-Allow-Origin": "https://myapp.com",
-   *   "Access-Control-Allow-Credentials": "true",
-   *   "Access-Control-Allow-Methods": "GET, POST, HEAD, OPTIONS",
-   *   "Access-Control-Allow-Headers": "Content-Type, Authorization"
-   * }
-   * ```
-   *
-   * When set to a `HeadersInit` value, uses those as the CORS headers instead.
-   * CORS preflight (OPTIONS) requests are handled automatically for matched routes.
-   * Non-WebSocket responses on matched routes will also have the CORS headers appended.
-   */
-  cors?: boolean | HeadersInit;
-  /**
-   * Retry transient Durable Object infrastructure errors thrown while routing
-   * to the target DO. Enabled by default; pass `false` to disable.
-   *
-   * Only errors marked `retryable === true` are retried, and overloaded
-   * errors (`overloaded === true`) are never retried.
-   */
-  routingRetry?: false | RoutingRetryOptions;
-  onBeforeConnect?: (
-    req: Request,
-    route: DurableObjectRouteMatch<Env>
-  ) => Response | Request | void | Promise<Response | Request | void>;
-  onBeforeRequest?: (
-    req: Request,
-    route: DurableObjectRouteMatch<Env>
-  ) =>
-    | Response
-    | Request
-    | void
-    | Promise<Response | Request | undefined | void>;
-}
-/**
- * Resolve CORS options into a concrete headers object (or null if CORS is disabled).
- */
-function resolveCorsHeaders(
-  cors: boolean | HeadersInit | undefined
-): Record<string, string> | null {
-  if (cors === true) {
-    return {
-      "Access-Control-Allow-Origin": "*",
-      "Access-Control-Allow-Methods": "GET, POST, HEAD, OPTIONS",
-      "Access-Control-Allow-Headers": "*",
-      "Access-Control-Max-Age": "86400"
-    };
-  }
-  if (cors && typeof cors === "object") {
-    // Convert any HeadersInit shape to a plain record
-    const h = new Headers(cors as HeadersInit);
-    const record: Record<string, string> = {};
-    h.forEach((value, key) => {
-      record[key] = value;
-    });
-    return record;
-  }
-  return null;
-}
-
-/**
- * Route `/objects/:binding/:name` requests to named Durable Objects.
- *
- * @param req - Incoming Worker request.
- * @param env - Worker environment containing Durable Object bindings.
- * @param options - Route hooks, prefix, placement, CORS, props, and retries.
- * @returns The matched response, or `null` when the path does not match.
- */
-export async function routeDurableObjectRequest<
-  Env extends object = Cloudflare.Env,
-  T extends DurableObject<Env> = DurableObject<Env>,
-  Props extends Record<string, unknown> = Record<string, unknown>
->(
-  req: Request,
-  env: Env = defaultEnv as Env,
-  options?: DurableObjectRouteOptions<Env, Props>
-): Promise<Response | null> {
-  if (!namespaceMapCache.has(env)) {
-    const namespaceMap: Record<string, DurableObjectNamespace> = {};
-    const bindingNames: Record<string, string> = {};
-    for (const [k, v] of Object.entries(env)) {
-      if (
-        v &&
-        typeof v === "object" &&
-        "idFromName" in v &&
-        typeof v.idFromName === "function"
-      ) {
-        const kebab = camelCaseToKebabCase(k);
-        namespaceMap[kebab] = v as DurableObjectNamespace;
-        bindingNames[kebab] = k;
-      }
-    }
-    namespaceMapCache.set(env, namespaceMap);
-    bindingNameCache.set(env, bindingNames);
-  }
-  const map = namespaceMapCache.get(env) as unknown as Record<
-    string,
-    DurableObjectNamespace<T>
-  >;
-  const bindingNames = bindingNameCache.get(env) as Record<string, string>;
-
-  const prefix = options?.prefix || "objects";
-  const prefixParts = prefix.split("/");
-
-  const url = new URL(req.url);
-  const parts = url.pathname.split("/").filter(Boolean); // Remove empty strings
-
-  // Check if the URL starts with the prefix
-  const prefixMatches = prefixParts.every(
-    (part, index) => parts[index] === part
-  );
-  if (!prefixMatches || parts.length < prefixParts.length + 2) {
-    return null;
-  }
-
-  const namespace = parts[prefixParts.length];
-  const name = parts[prefixParts.length + 1];
-
-  if (name && namespace) {
-    if (!map[namespace]) {
-      console.error(
-        `The URL ${req.url} with namespace "${namespace}" and name "${name}" does not match any Durable Object binding.`
-      );
-      // we should return a response with a status code that it's an invalid request
-      return new Response("Invalid request", { status: 400 });
-    }
-
-    // Resolve CORS headers for this matched route
-    const corsHeaders = resolveCorsHeaders(options?.cors);
-    const isWebSocket =
-      req.headers.get("Upgrade")?.toLowerCase() === "websocket";
-
-    // Helper: append CORS headers to a response (skipped for WebSocket upgrades)
-    function withCorsHeaders(response: Response): Response {
-      if (!corsHeaders || isWebSocket) return response;
-      const newResponse = new Response(response.body, response);
-      for (const [key, value] of Object.entries(corsHeaders)) {
-        newResponse.headers.set(key, value);
-      }
-      return newResponse;
-    }
-
-    // Handle CORS preflight requests for matched routes
-    if (req.method === "OPTIONS" && corsHeaders) {
-      return new Response(null, { headers: corsHeaders });
-    }
-
-    let doNamespace = map[namespace];
-    if (options?.jurisdiction) {
-      doNamespace = doNamespace.jurisdiction(options.jurisdiction);
-    }
-
-    const id = doNamespace.idFromName(name);
-    const getOptions = durableObjectGetOptions(options);
-
-    req = mutableRequest(req);
-
-    const className = bindingNames[namespace] as Extract<keyof Env, string>;
-    const route: DurableObjectRouteMatch<Env> = { className, name };
-
-    if (isWebSocket) {
-      if (options?.onBeforeConnect) {
-        const reqOrRes = await options.onBeforeConnect(req, route);
-        if (reqOrRes instanceof Request) {
-          req = mutableRequest(reqOrRes);
-        } else if (reqOrRes instanceof Response) {
-          return reqOrRes;
-        }
-      }
-    } else {
-      if (options?.onBeforeRequest) {
-        const reqOrRes = await options.onBeforeRequest(req, route);
-        if (reqOrRes instanceof Request) {
-          req = mutableRequest(reqOrRes);
-        } else if (reqOrRes instanceof Response) {
-          return withCorsHeaders(reqOrRes);
-        }
-      }
-    }
-
-    // Attach props to the request after the hooks so that user-defined
-    // onBeforeConnect / onBeforeRequest callbacks don't see the serialized
-    // props header on the inspection request.
-    if (options?.props !== undefined) {
-      req.headers.set("x-agents-lifecycle-props", encodeProps(options.props));
-    }
-
-    // Props travel with the request and are applied before startup.
-    const response = await retryDurableObjectOperation(
-      () => doNamespace.get(id, getOptions).fetch(cloneRequestForFetch(req)),
-      { name, className },
-      options?.routingRetry
-    );
-    return isWebSocket ? response : withCorsHeaders(response);
-  } else {
-    return null;
-  }
 }
 
 type LifecycleHost<Props extends object> = {
