+++ packages/agents/src/lifecycle/index.ts
@@ -1,17 +1,12 @@
 /** Durable Object lifecycle composition for Agents and standalone objects. */
 export {
   Lifecycle,
-  routeDurableObjectRequest,
   type Connection,
   type ConnectionContext,
   type ConnectionSetStateFn,
   type ConnectionState,
   type DurableObjectCapability,
   type CapabilityRequestContext,
-  type DurableObjectRouteMatch,
-  type DurableObjectRouteOptions,
   type CapabilityStartContext,
-  type RoutingRetryEvent,
-  type RoutingRetryOptions,
   type WSMessage
 } from "./durable-object-lifecycle";
